+++ src/app/api/crud/followings/route.ts
@@ -1,7 +1,6 @@
 import { Following } from "@/lib/db/models";
 import pool from "../../../../lib/db/pool";
-import getCurrentSession from "@/lib/cookies/getCurrentSession";
-import { NextResponse } from "next/server";
+import getCurrentSessionUserId from "@/lib/cookies/getCurrentSessionUserId";
 
 /**
  * GET endpoint for table followings (Fetch all followings)
@@ -41,8 +40,11 @@ export async function GET(request: Request): Promise<Response> {
  */
 export async function POST(request: Request): Promise<Response> {
     let client;
-    const userId = await getCurrentSession();
     try {
+        const userId = await getCurrentSessionUserId();
+        if (userId === -1) {
+            return new Response("Unauthorized API call", { status: 401 });
+        }
         const following: Following = await request.json();
         following.userId = userId;
         client = await pool.connect();
@@ -71,8 +73,11 @@ export async function POST(request: Request): Promise<Response> {
  */
 export async function PUT(request: Request): Promise<Response> {
     let client;
-    const userId = await getCurrentSession();
     try {
+        const userId = await getCurrentSessionUserId();
+        if (userId === -1) {
+            return new Response("Unauthorized API call", { status: 401 });
+        }
         const following: Following = await request.json();
         following.userId = userId;
         client = await pool.connect();
@@ -103,6 +108,10 @@ export async function DELETE(request: Request): Promise<Response> {
     let client;
     try {
         const { id } = await request.json();
+        const userId = await getCurrentSessionUserId();
+        if (userId === -1 || userId !== id) {
+            return new Response("Unauthorized API call", { status: 401 });
+        }
         client = await pool.connect();
         await client.query("DELETE FROM followings WHERE following_status_id = $1", [id]);
