+++ src/app/api/crud/posts/route.ts
@@ -1,8 +1,6 @@
 import { Post } from "@/lib/db/models";
 import pool from "@/lib/db/pool";
-import getCurrentSession from "@/lib/cookies/getCurrentSession";
-import { redirect } from "next/navigation";
-import { NextResponse } from "next/server";
+import getCurrentSessionUserId from "@/lib/cookies/getCurrentSessionUserId";
 
 /**
  * GET endpoint for table posts (Fetch all posts)
@@ -43,8 +41,11 @@ export async function GET(request: Request): Promise<Response> {
  */
 export async function POST(request: Request): Promise<Response> {
     let client;
-    const userId = await getCurrentSession();
     try {
+        const userId = await getCurrentSessionUserId();
+        if (userId === -1) {
+            return new Response("Unauthorized API call", { status: 401 });
+        }
         const post: Post = await request.json();
         post.userId = userId;
         client = await pool.connect();
@@ -74,8 +75,11 @@ export async function POST(request: Request): Promise<Response> {
  */
 export async function PUT(request: Request): Promise<Response> {
     let client;
-    const userId = await getCurrentSession();
     try {
+        const userId = await getCurrentSessionUserId();
+        if (userId === -1) {
+            return new Response("Unauthorized API call", { status: 401 });
+        }
         const post: Post = await request.json();
         post.userId = userId;
         client = await pool.connect();
@@ -106,6 +110,10 @@ export async function DELETE(request: Request): Promise<Response> {
     let client;
     try {
         const { id } = await request.json();
+        const userId = await getCurrentSessionUserId();
+        if (userId === -1 || userId !== id) {
+            return new Response("Unauthorized API call", { status: 401 });
+        }
         client = await pool.connect();
         await client.query("DELETE FROM posts WHERE post_id = $1", [id]);
