+++ src/app/api/crud/user_settings/route.ts
@@ -1,7 +1,6 @@
 import { UserSettings } from "@/lib/db/models";
 import pool from "../../../../lib/db/pool";
-import getCurrentSession from "@/lib/cookies/getCurrentSession";
-import { NextResponse } from "next/server";
+import getCurrentSessionUserId from "@/lib/cookies/getCurrentSessionUserId";
 
 /**
  * GET endpoint for table user_settings (Fetch all user settings)
@@ -42,8 +41,11 @@ export async function GET(request: Request): Promise<Response> {
  */
 export async function POST(request: Request): Promise<Response> {
     let client;
-    const userId = await getCurrentSession();
     try {
+        const userId = await getCurrentSessionUserId();
+        if (userId === -1) {
+            return new Response("Unauthorized API call", { status: 401 });
+        }
         const userSettings: UserSettings = await request.json();
         userSettings.userId = userId;
         client = await pool.connect();
@@ -73,9 +75,9 @@ export async function POST(request: Request): Promise<Response> {
 export async function PUT(request: Request): Promise<Response> {
     let client;
     try {
-        const userId = await getCurrentSession();
+        const userId = await getCurrentSessionUserId();
         if (userId === -1) {
-            return NextResponse.redirect(new URL("/session_inactive", request.url));
+            return new Response("Unauthorized API call", { status: 401 });
         }
         const userSettings: UserSettings = await request.json();
         userSettings.userId = userId;
@@ -107,6 +109,10 @@ export async function DELETE(request: Request): Promise<Response> {
     let client;
     try {
         const { id } = await request.json();
+        const userId = await getCurrentSessionUserId();
+        if (userId === -1 || userId !== id) {
+            return new Response("Unauthorized API call", { status: 401 });
+        }
         client = await pool.connect();
         await client.query("DELETE FROM user_settings WHERE user_settings_id = $1", [id]);
